fix(LLParser): fix bugs associated with synthesized and action records.
Synthesized and Action records are now fully supported.

The records' actions are executed in a preorder traversal.
The action gets access to the parser stack and the data
record itself.
The data record must be of trivial type (this may change
in a later release).
You must cast the type of the action function to `void*` when
constructing the record.