feat(ItemSet): finish the implementation of CLOSURE.
This commit finishes the implementation of the CLOSURE
function.

It also solves a bug in `Item` string conversion.

Precise size calculation is added to `ProdVec`; it
considers only grammar symbols that form part of the
production body.
Similar functionality has been added to the dot
position of an item to get the actual dot position
within the production of the item.