test(LRParser): enhance LR parser test (calculator).
This commit tests action support for the parser.
It enhances the test to support actions, particularly
so that it is now a fully-functioning sub-calculator.
To support actions, LRState must store the current
token when the state is shifted, so a new member has
been added to LRState.
LRState conversion to string is also now supported.

This commit also adds logging to the parser so that the
parser now logs its actions.

New conversion function (string to int) has been added
to common.h.