feat(LRPGenerator): finish LALR(1) parser generator.
This commit adds LALR(1) parser generator.
Several utility functions were added to `ItemSet`
It also refactors the existing implementations of
parser generators and reuses the refactored functions
in the creation of this new parser generator.
It is noteworthy, however, that the algorithm used is
not the most optimal algorithm for that type of
parser generator.