feat(regex)!: create the grammar and basic structures.
This commit creates the grammar of the regular expression
, However, it does not add the semantic actiosn that
will perform the translation.

BREAKING_CHANGE It also adds a means for resolving conflicts.
When a conflict occurs, now an exception will not be
thrown, rather the user will be presented with
information about the conflict and will prompted to
resolve it.