feat(regex)!: finish parsing pattern and `match`.
- This commit finishes parsing the pattern given.
It also implements the `match` function, although
it lacks functionality that need handling flags.

- There are also certain bug fixes, especially when
it comes to parsing the pattern.

- A `Test.cpp` file has been added containing syntax
from GTest. Testing with GTest will be added by the
next commit.

BREAKING_CHANGE The result returned by the parser
on calling `parse` is now specified by a template
argument.